pyproject.toml
[tool.poetry]
name = "loopr"
version = "0.1.0"
description = "A python sdk to interact with loopR APIs."
authors = ["LoopR"]

[tool.poetry.dependencies]
python = "^3.6"
loguru = "^0.5.3"
python-dotenv = "^0.14.0"
requests = "^2.24.0"
python-dateutil = "^2.8.1"

[tool.poetry.dev-dependencies]
isort = "^5.6.4"
black = "^20.8b1"
autoflake = "^1.4"
mypy = "^0.790"
twine = "^3.2.0"
coverage = "^5.3"
flake8 = "^3.8.4"
pytest = "^6.1.1"
pytest-cov = "^2.10.1"
Sphinx = "^3.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
line_length = 88
force_grid_wrap = 0
combine_as_imports = true


[tool.black]
line-length = 88
include = '\.pyi?$'